chore: move factory function to own client.py
